+++ apps/todos/src/app/features/todos/containers/todos/todos.component.html
@@ -1,4 +1,4 @@
-<learning-workspace-search></learning-workspace-search>
+<learning-workspace-search (search)="searchHandle($event)"></learning-workspace-search>
 
 <ul *ngIf="todos?.length" class="list-group">  
   <ng-container *ngFor="let todo of todos; trackBy:todoIdentify">
