+++ apps/todos/src/app/modules/about/about-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Route } from '@angular/router';
 
-import { AboutComponent } from './containers';
+import { AboutComponent } from './about.component';
 
 const routes: Route[] = [
   {
