--- apps/todos/src/app/modules/about/containers/about/index.ts
@@ -1,1 +1,0 @@
-export * from './about.component'
