--- apps/todos/src/app/modules/about/containers/index.ts
@@ -1,1 +1,0 @@
-export * from './about';
