--- apps/todos/src/app/modules/add-edit-todo-form/container/index.ts
@@ -1,1 +1,0 @@
-export * from './add-edit-todo-form'
