--- apps/todos/src/app/modules/footer/container/footer/index.ts
@@ -1,1 +1,0 @@
-export * from './footer.component'
