--- apps/todos/src/app/modules/footer/container/index.ts
@@ -1,1 +1,0 @@
-export * from './footer'
