+++ apps/todos/src/app/modules/footer/footer.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FooterComponent } from './container';
+import { FooterComponent } from './footer.component';
 
 @NgModule({
   declarations: [
