+++ apps/todos/src/app/modules/header/header.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 
-import { HeaderComponent } from './container/header/header.component';
+import { HeaderComponent } from './header.component';
 
 @NgModule({
   declarations: [
