--- apps/todos/src/app/modules/page-not-found/container/index.ts
@@ -1,1 +1,0 @@
-export * from './page-not-found'
