--- apps/todos/src/app/modules/page-not-found/container/page-not-found/index.ts
@@ -1,1 +1,0 @@
-export * from './page-not-found.component'
