+++ apps/todos/src/app/modules/page-not-found/page-not-found.module.ts
@@ -4,7 +4,7 @@ import { RouterModule } from '@angular/router';
 
 
 import { PageNotFoundRouterModule } from './page-not-found-routing.module';
-import { PageNotFoundComponent } from './container';
+import { PageNotFoundComponent } from './page-not-found.component';
 
 @NgModule({
   declarations: [
