--- apps/todos/src/app/modules/search/container/index.ts
@@ -1,1 +1,0 @@
-export * from './search'
