+++ apps/todos/src/app/modules/search/search.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 
-import { SearchComponent } from './container';
+import { SearchComponent } from './search.component';
